Fix for macOS
