+++ README.md
@@ -103,5 +103,3 @@ view.topSheet(
     presentOn: { UIViewController() }
 )
 ```
-
-
